Cleanup for latest Lilu

+++ AirportBrcmFixup/kern_fakebrcm.cpp
@@ -1,14 +1,15 @@
-#include <Library/LegacyIOService.h>
-#include <Headers/kern_iokit.hpp>
+//
+//  kern_fakebrcm.cpp
+//  AirportBrcmFixup
+//
+//  Copyright © 2017 lvs1974. All rights reserved.
+//
+
 #include <Headers/plugin_start.hpp>
-//#include <Headers/kern_api.hpp>
-#include <Headers/kern_patcher.hpp>
 
 #include "kern_config.hpp"
 #include "kern_fakebrcm.hpp"
 #include "kern_misc.hpp"
-
-
 
 OSDefineMetaClassAndStructors(FakeBrcm, IOService);
 
@@ -22,22 +23,22 @@ WIOKit::t_PCIConfigRead32  FakeBrcm::orgConfigRead32 {nullptr};
 
 bool FakeBrcm::init(OSDictionary *propTable)
 {
-    if (ADDPR(brcmfx_config).disabled)
-    {
-        DBGLOG("BRCMFX", "FakeBrcm::init(): FakeBrcm disabled");
-        return false;
-    }
-    
-    DBGLOG("BRCMFX", "FakeBrcm::init()");
+	if (ADDPR(brcmfx_config).disabled)
+	{
+		DBGLOG("BRCMFX", "FakeBrcm::init(): FakeBrcm disabled");
+		return false;
+	}
+	
+	DBGLOG("BRCMFX", "FakeBrcm::init()");
 
-    bool ret = super::init(propTable);
-    if (!ret)
-    {
-        SYSLOG("BRCMFX", "FakeBrcm super::init returned false\n");
-        return false;
-    }
-    
-    return true;
+	bool ret = super::init(propTable);
+	if (!ret)
+	{
+		SYSLOG("BRCMFX", "FakeBrcm super::init returned false\n");
+		return false;
+	}
+	
+	return true;
 }
 
 //==============================================================================
@@ -45,69 +46,69 @@ bool FakeBrcm::init(OSDictionary *propTable)
 bool FakeBrcm::attach(IOService *provider)
 {
 	service_provider = provider;
-    hookProvider(provider);
-    
-    return super::attach(provider);
+	hookProvider(provider);
+	
+	return super::attach(provider);
 }
 
 //==============================================================================
 
 IOService* FakeBrcm::probe(IOService * provider, SInt32 *score)
 {
-    DBGLOG("BRCMFX", "FakeBrcm::probe()");
+	DBGLOG("BRCMFX", "FakeBrcm::probe()");
 
 	if (!ADDPR(startSuccess))
 	{
 		return nullptr;
 	}
 
-    IOService* ret = super::probe(provider, score);
-    if (!ret)
-    {
-        SYSLOG("BRCMFX", "FakeBrcm super::probe returned nullptr\n");
-        return nullptr;
-    }
+	IOService* ret = super::probe(provider, score);
+	if (!ret)
+	{
+		SYSLOG("BRCMFX", "FakeBrcm super::probe returned nullptr\n");
+		return nullptr;
+	}
 	
 	*score = -2000;
 
-    if (ADDPR(brcmfx_config).disabled)
-    {
-        DBGLOG("BRCMFX", "FakeBrcm::probe(): FakeBrcm disabled");
-        return ret;
-    }
+	if (ADDPR(brcmfx_config).disabled)
+	{
+		DBGLOG("BRCMFX", "FakeBrcm::probe(): FakeBrcm disabled");
+		return ret;
+	}
 	
-    DBGLOG("BRCMFX", "FakeBrcm::probe(): service provider is %s", provider->getName());
-    
-    for (int i=0; i<kextListSize; i++)
-    {
-        const OSMetaClass * meta_class = OSMetaClass::getMetaClassWithName(OSSymbol::withCStringNoCopy(serviceNameList[i]));
-        if (meta_class != nullptr)
-        {
-            DBGLOG("BRCMFX", "FakeBrcm::probe(): meta class for %s exists!", serviceNameList[i]);
+	DBGLOG("BRCMFX", "FakeBrcm::probe(): service provider is %s", provider->getName());
+	
+	for (size_t i = 0; i < kextListSize; i++)
+	{
+		const OSMetaClass * meta_class = OSMetaClass::getMetaClassWithName(OSSymbol::withCStringNoCopy(serviceNameList[i]));
+		if (meta_class != nullptr)
+		{
+			DBGLOG("BRCMFX", "FakeBrcm::probe(): meta class for %s exists!", serviceNameList[i]);
 
-            IOService *service = (IOService *) OSMetaClass::allocClassWithName(serviceNameList[i]);
-            if (service != nullptr)
-            {
-                service->retain();
-                DBGLOG("BRCMFX", "FakeBrcm: retain counter for driver %s is %d", serviceNameList[i], service->getRetainCount());
+			IOService *service = (IOService *) OSMetaClass::allocClassWithName(serviceNameList[i]);
+			if (service != nullptr)
+			{
+				service->retain();
+				DBGLOG("BRCMFX", "FakeBrcm: retain counter for driver %s is %d", serviceNameList[i], service->getRetainCount());
 				service_found = true;
-            }
-            else
-                SYSLOG("BRCMFX", "FakeBrcm: instance of driver %s couldn't be created", serviceNameList[i]);
-        }
-    }
+			}
+			else
+				SYSLOG("BRCMFX", "FakeBrcm: instance of driver %s couldn't be created", serviceNameList[i]);
+		}
+	}
 	
-    if (service_found)
-    {
-        DBGLOG("BRCMFX", "FakeBrcm::probe() will change score from %d to 2000", *score);
-        *score = 2000;  // change probe score to be the first in the list
-    }
-    else
-    {
-        DBGLOG("BRCMFX", "FakeBrcm::probe(): fallback to original driver");
-    }
+	if (service_found)
+	{
+		DBGLOG("BRCMFX", "FakeBrcm::probe() will change score from %d to 2000", *score);
+		*score = 2000;  // change probe score to be the first in the list
+	}
+	else
+	{
+		DBGLOG("BRCMFX", "FakeBrcm::probe(): fallback to original driver");
+	}
 
-    return ret;
+	return ret;
 }
 
 //==============================================================================
@@ -121,44 +122,44 @@ bool FakeBrcm::start(IOService *provider)
 		return nullptr;
 	}
 
-    if (ADDPR(brcmfx_config).disabled)
-    {
-        DBGLOG("BRCMFX", "FakeBrcm::start(): FakeBrcm disabled");
-        return false;
-    }
-    
-    if (!super::start(provider))
-    {
-        SYSLOG("BRCMFX", "FakeBrcm super::start returned false\n");
-        return false;
-    }
-    
-    if (!service_found)
-    {
-        SYSLOG("BRCMFX", "FakeBrcm::start(): fallback to original driver");
-        return false;
-    }
+	if (ADDPR(brcmfx_config).disabled)
+	{
+		DBGLOG("BRCMFX", "FakeBrcm::start(): FakeBrcm disabled");
+		return false;
+	}
+	
+	if (!super::start(provider))
+	{
+		SYSLOG("BRCMFX", "FakeBrcm super::start returned false\n");
+		return false;
+	}
+	
+	if (!service_found)
+	{
+		SYSLOG("BRCMFX", "FakeBrcm::start(): fallback to original driver");
+		return false;
+	}
 	
 	service_provider = provider;
-    hookProvider(provider);
+	hookProvider(provider);
 
-    return true;
+	return true;
 }
 
 //==============================================================================
 
 void FakeBrcm::stop(IOService *provider)
 {
-    DBGLOG("BRCMFX", "FakeBrcm::stop()");
-    super::stop(provider);
+	DBGLOG("BRCMFX", "FakeBrcm::stop()");
+	super::stop(provider);
 }
 
 //==============================================================================
 
 void FakeBrcm::free()
 {
-    DBGLOG("BRCMFX", "FakeBrcm::free()");
-    super::free();
+	DBGLOG("BRCMFX", "FakeBrcm::free()");
+	super::free();
 }
 
 
@@ -181,43 +182,43 @@ bool FakeBrcm::isServiceSupported(IORegistryEntry* service)
 
 UInt16 FakeBrcm::configRead16(IORegistryEntry *service, UInt32 space, UInt8 offset)
 {
-    UInt16 result = orgConfigRead16(service, space, offset);
-    UInt16 newResult = result;
-    
-    if ((offset == WIOKit::PCIRegister::kIOPCIConfigVendorID || offset == WIOKit::PCIRegister::kIOPCIConfigDeviceID) && isServiceSupported(service))
-    switch (offset)
-    {
-        case WIOKit::PCIRegister::kIOPCIConfigVendorID:
-        {
-            UInt32 vendor;
-            if (WIOKit::getOSDataValue(service, "vendor-id", vendor))
-                newResult = vendor;
-            break;
-        }
-        case WIOKit::PCIRegister::kIOPCIConfigDeviceID:
-        {
-            UInt32 device;
-            if (WIOKit::getOSDataValue(service, "device-id", device))
-                newResult = device;
-            break;
-        }
-    }
-    
-    if (newResult != result)
-        DBGLOG("BRCMFX", "FakeBrcm::configRead16: name = %s, source value = 0x%04x replaced with value = 0x%04x", service->getName(), result, newResult);
+	UInt16 result = orgConfigRead16(service, space, offset);
+	UInt16 newResult = result;
+	
+	if ((offset == WIOKit::PCIRegister::kIOPCIConfigVendorID || offset == WIOKit::PCIRegister::kIOPCIConfigDeviceID) && isServiceSupported(service))
+	switch (offset)
+	{
+		case WIOKit::PCIRegister::kIOPCIConfigVendorID:
+		{
+			UInt32 vendor;
+			if (WIOKit::getOSDataValue(service, "vendor-id", vendor))
+				newResult = vendor;
+			break;
+		}
+		case WIOKit::PCIRegister::kIOPCIConfigDeviceID:
+		{
+			UInt32 device;
+			if (WIOKit::getOSDataValue(service, "device-id", device))
+				newResult = device;
+			break;
+		}
+	}
+	
+	if (newResult != result)
+		DBGLOG("BRCMFX", "FakeBrcm::configRead16: name = %s, source value = 0x%04x replaced with value = 0x%04x", service->getName(), result, newResult);
 
-    return newResult;
+	return newResult;
 }
 
 //==============================================================================
 
 UInt32 FakeBrcm::configRead32(IORegistryEntry *service, UInt32 space, UInt8 offset)
 {
-    UInt32 result = orgConfigRead32(service, space, offset);
-    UInt32 newResult = result;
+	UInt32 result = orgConfigRead32(service, space, offset);
+	UInt32 newResult = result;
 	
 	// OS X does a non-aligned read, which still returns full vendor / device ID
-    if ((offset == WIOKit::PCIRegister::kIOPCIConfigVendorID || offset == WIOKit::PCIRegister::kIOPCIConfigDeviceID) && isServiceSupported(service))
+	if ((offset == WIOKit::PCIRegister::kIOPCIConfigVendorID || offset == WIOKit::PCIRegister::kIOPCIConfigDeviceID) && isServiceSupported(service))
 	{
 		UInt32 vendor, device;
 		if (WIOKit::getOSDataValue(service, "vendor-id", vendor))
@@ -225,11 +226,11 @@ UInt32 FakeBrcm::configRead32(IORegistryEntry *service, UInt32 space, UInt8 offs
 		if (WIOKit::getOSDataValue(service, "device-id", device))
 			newResult = (device << 16) | (newResult & 0xFFFF);
 	}
-    
-    if (newResult != result)
-        DBGLOG("BRCMFX", "FakeBrcm::configRead32: name = %s, source value = 0x%08x replaced with value = 0x%08x", service->getName(), result, newResult);
-    
-    return newResult;
+	
+	if (newResult != result)
+		DBGLOG("BRCMFX", "FakeBrcm::configRead32: name = %s, source value = 0x%08x replaced with value = 0x%08x", service->getName(), result, newResult);
+	
+	return newResult;
 }
 
 //==============================================================================
@@ -255,4 +256,3 @@ void FakeBrcm::hookProvider(IOService *provider)
 		if (KernelPatcher::routeVirtual(provider, WIOKit::PCIConfigOffset::ConfigRead32, configRead32, &orgConfigRead32))
 			DBGLOG("BRCMFX", "FakeBrcm::hookProvider for configRead32 was successful");
 }
-
